# openapi.yaml
openapi: 3.0.1
info:
  description: API of EtuEDT-Back server.
  version: 2.2.0
  title: EtuEDT-Back
servers:
  - url: https://edtapi.antoninhuaut.fr/v2
paths:
  /:
    get:
      tags:
        - general
      summary: Get the base routes
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                    example: v2
  /univ:
    get:
      tags:
        - university
      summary: Get the list of universities
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/University'
  /univ/{univId}:
    get:
      tags:
        - university
      summary: Get university detail
      parameters:
        - name: univId
          in: path
          description: Id of university
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/University'
        '404':
          description: University not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /univ/{univId}/groups:
    get:
      tags:
        - group
      summary: Get the list of groups for a university
      parameters:
        - name: univId
          in: path
          description: Id of university
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '404':
          description: University not found
  /univ/{univId}/groups/{groupId}:
    get:
      tags:
        - group
      summary: Get the list of timetables in a group
      parameters:
        - name: univId
          in: path
          description: Id of university
          required: true
          schema:
            type: integer
        - name: groupId
          in: path
          description: Id of group
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Timetable'
        '404':
          description: University or group not found
  /univ/{univId}/groups/{groupId}/{adeResources}:
    get:
      tags:
        - timetable
      summary: Get timetable metadata
      parameters:
        - name: univId
          in: path
          description: Id of university
          required: true
          schema:
            type: integer
        - name: groupId
          in: path
          description: Id of group
          required: true
          schema:
            type: integer
        - name: adeResources
          in: path
          description: ADE resources ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timetable'
        '404':
          description: Timetable not found
  /univ/{univId}/groups/{groupId}/{adeResources}/events:
    get:
      tags:
        - timetable
      summary: Get timetable events
      parameters:
        - name: univId
          in: path
          description: Id of university
          required: true
          schema:
            type: integer
        - name: groupId
          in: path
          description: Id of group
          required: true
          schema:
            type: integer
        - name: adeResources
          in: path
          description: ADE resources ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JsonEvent'
            text/calendar:
              schema:
                $ref: '#/components/schemas/IcalEvent'
        '503':
          description: No cache found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /univ/{univId}/rooms:
    get:
      tags:
        - room
      summary: Get the list of rooms for a university
      parameters:
        - name: univId
          in: path
          description: Id of university
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '404':
          description: University not found
  /univ/{univId}/rooms/{adeResources}:
    get:
      tags:
        - room
      summary: Get room metadata
      parameters:
        - name: univId
          in: path
          description: Id of university
          required: true
          schema:
            type: integer
        - name: adeResources
          in: path
          description: ADE resources ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          description: Room not found
  /univ/{univId}/rooms/{adeResources}/events:
    get:
      tags:
        - room
      summary: Get room events
      parameters:
        - name: univId
          in: path
          description: Id of university
          required: true
          schema:
            type: integer
        - name: adeResources
          in: path
          description: ADE resources ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JsonEvent'
            text/calendar:
              schema:
                $ref: '#/components/schemas/IcalEvent'
        '503':
          description: No cache found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    University:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Université de Caen
    Group:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: BUT Informatique - 1ère année
    Timetable:
      type: object
      properties:
        adeResources:
          type: integer
          example: 1177
        adeProjectId:
          type: integer
          example: 2
        year:
          type: integer
          example: 1
        label:
          type: string
          example: TP 1.1
        adeUrl:
          type: string
          example: https://ade.unicaen.fr/...
        lastUpdate:
          type: string
          example: '2020-06-16T07:14:34.017Z'
    Room:
      type: object
      properties:
        adeResources:
          type: integer
          example: 42349
        adeProjectId:
          type: integer
          example: 2
        label:
          type: string
          example: AC 007 (A)
        adeUrl:
          type: string
          example: https://ade.unicaen.fr/...
        lastUpdate:
          type: string
          example: '2020-06-16T07:14:34.017Z'
    Error:
      type: object
      properties:
        error:
          type: string
          example: no cache found for this resource, try again later
    JsonEvent:
      type: object
      properties:
        title:
          type: string
          example: Reunion de rentrée 1A
        teacher:
          type: string
          example: M. Dupont
        description:
          type: string
          example: BUT 1A
        location:
          type: string
          example: Amphithéâtre
        start:
          type: string
          example: '2023-09-05T09:00:00Z'
        end:
          type: string
          example: '2023-09-05T10:00:00Z'
    IcalEvent:
      type: string
      example: |
        BEGIN:VCALENDAR
        METHOD:REQUEST
        PRODID:-//ADE/version 6.0
        VERSION:2.0
        CALSCALE:GREGORIAN
        BEGIN:VEVENT
        DTSTAMP:20231208T153544Z
        DTSTART:20230905T090000Z
        DTEND:20230905T100000Z
        SUMMARY:Reunion de rentrée 1A
        LOCATION:Amphithéâtre
        DESCRIPTION:\n\nBUT 1A\nM. Dupont\n(Exported :08/12/2023 16:35)\n
        UID:ADE60323032332d323032342d3135343639352d302d30
        CREATED:19700101T000000Z
        LAST-MODIFIED:20231208T153544Z
        SEQUENCE:-1960707552
        END:VEVENT
        END:VCALENDAR